safety commit
