+++ src/components/victory-events/victory-events.js
@@ -1,4 +1,4 @@
-import { assign, keys } from "lodash";
+import { assign, isFunction, partialRight } from "lodash";
 import React, { PropTypes } from "react";
 import { PropTypes as CustomPropTypes } from "victory-core";
 import Events from "../../helpers/events";
@@ -30,17 +30,62 @@ export default class VictoryEvents extends React.Component {
     this.getEventState = Events.getEventState.bind(this);
   }
 
-  getNewChildren(props) {
+  componentWillMount() {
+    this.setUpChildren(this.props);
+  }
+
+  componentWillReceiveProps(newProps) {
+    this.setUpChildren(newProps);
+  }
+
+  setUpChildren(props) {
     const childComponents = React.Children.toArray(props.children);
+    this.baseProps = this.getBasePropsFromChildren(childComponents);
+    this.newChildren = this.getNewChildren(props, childComponents);
+  }
+
+  getBasePropsFromChildren(childComponents) {
+    const childTypes = [];
+    const getChildKey = (child) => {
+      if (child.props.name) {
+        return child.props.name;
+      }
+      const role = child.type && child.type.role;
+      const count = childTypes.filter((type) => type === role).length;
+      return count ? `${role}-${count}` : role;
+    };
+
+    const getBaseProps = (children) => {
+      return children.reduce((memo, child) => {
+        if (child.type && isFunction(child.type.getBaseProps)) {
+          const baseChildProps = child.props && child.type.getBaseProps(child.props);
+          if (baseChildProps) {
+            const childKey = getChildKey(child);
+            memo[childKey] = baseChildProps;
+            childTypes.push(child.type.role);
+            return memo;
+          }
+          return memo;
+        } else if (child.props && child.props.children) {
+          return getBaseProps(React.Children.toArray(child.props.children));
+        }
+        return memo;
+      }, {});
+    };
+    return getBaseProps(childComponents);
+  }
+
+  getNewChildren(props, childComponents) {
     const {events, eventKey} = props;
+    const childNames = Object.keys(this.baseProps);
     return childComponents.map((child, index) => {
       return React.cloneElement(child, assign(
         {
           key: `events-${index}`,
           sharedEvents: {
             events,
-            getEvents: this.getEvents,
-            getEventState: this.getEventState
+            getEvents: partialRight(this.getEvents, childNames[index]),
+            getEventState: partialRight(this.getEventState, childNames[index])
           },
           eventKey
         },
@@ -50,7 +95,7 @@ export default class VictoryEvents extends React.Component {
   }
 
   render() {
-    return <g>{this.getNewChildren(this.props)}</g>;
+    return <g>{this.newChildren}</g>;
   }
 
 }
